Pygame-RPG 15.1 Housekeeping and bugfixing
This commit fixes the save files, Rpg2.py file as well as the changing
the color for the health and mana bar text.

The save files were outdated due to the addition of the addition of
the `Mp` and `Mpcap` attributes being added to the Knight object. As a
result, this broke the old save files. This escaped automated testing
because the Dummy_Knight.py file did not have these changes. A fix for
Dummy_Knight.py will happen in the future. I also took the time to
change up the health and mana values to see what they would look like on
screen.

Rpg2.py now needed the addition of the BattleManager class since the
UIHandler class now needs it in it's constructor. With this addition,
the file can now run.

The color for the health and mana bar text used to be black but after
changing the save files, I realized that, on darker backgrounds it's
harder to see. I have changed the color to be white for now to fix this.