Pygame-RPG 16 Integrating Battle System
This change includes adding code from my test environment for the battle
system into the Rpg2.py file. The changes, other than adding the test
environment code, include fixing a glitch with character and
interactable animations as well as changing the condition for overworld
movement from the player's status being normal to a battle not
happening. I also included a bit of code to access a mock battle
environment by pressing "b".

This battle system is still a prototype and needs a variety of
enhancements. These enhancements depend on other classes being upgraded
however but this PR will serve as a template for how the battle system
will work in the future.