Pygame-RPG 17.1 ItemManager Integration
This PR is for integrating ItemManager functions into the battle flow of
the game. Essentially, this is picking up all of the pieces of
ItemManager and fitting them into the `Rpg2.py` file. To accommodate this
change, functions were added and some behaviors were tweaked. The JSON
structure for the `Item_Effects.json` file was also changed to provide more
information regarding item effects.The classes changed are the following,
`BattleManager`, `Knight`, `ItemManager`, `UIHandler` and `UIManager`.

`BattleManager`: BattleManager now has a reference to the ItemManager
object and the `heroPos` attribute now has valid coordinates, these
coordinates are still subject to change. Some helper functions were
created to streamline the integration. These functions are
`get_enemy_positions()` and `get_enemy_objects()`.

`get_enemy_positions()`: returns a list of x, y coordinate tuples based
on the `enemies` attribute list.

`get_enemy_objects()`: returns a list of Objects of the Enemy type based
on the `enemies` attribute list.

The main change to this class is how the `do_one_turn()` function works.
A new variable called `refresh` is set to false at the start of the
function. This `refresh` is a boolean that indicates that the submenu
needs to be refreshed. Once the function receives a valid `move`
and `target` variable, the code checks if the move is really an item
by checking it against all item names from the referenced ItemManager.
If the move is indeed an item, `refresh` is set to True and the effects
of the item is retrieved. The function then calls on the new
`use_items()`. If the choice isn't an item, it is considered a move and
processed as such. `do_one_turn returns but the `returnable_strings`
array and the `refresh` boolean value.

`use_items()`: This function takes in the object that is using the item,
the name of the item, the itemEffect dictionary and the targets for the
item's effect. A function that processes the use of an item. It is
essentially a simplified, `use_moves()`. This returns a
`returnable_strings` array.

`Knight`: The Knight class received 2 helper functions,
`add_to_inventory()` and `remove_from_inventory()` as well as a
streamlining to the `get_rewards()` function.

`add_to_inventory()`: This function takes in an item string and an amount,
the amount is by default set to 1 unless a value is given. This function serves as a simplified ways to add to the player's inventory.

`remove_from_inventory()`: This function takes in an item string and an
amount, the amount is by default set to 1 unless a value is given. This
function serves as a streamlined way of removing items from the players
inventory.

`ItemManager`: This class had changes surrounding the structuring of how
item effects were to be proccessed. At the beginning, only the effect
string was stored and paired to the item's name in the
`Item_Effects.json` file. This however, was not efficient as it had too
little information to be properly processed. As such, it was decided
that the `Item_Effects.json` file would pair the item's name to a
dictionary that had the following keys, `Target`, `AOE` and `Effect`.

`Target`: is a string that indicates who the item can be used on. "S"
indicates that the item can only be used on the player character, "T"
indicates that the item can only be used on enemies and "All" means that
the target can be used on either.

`AOE`: is a boolean that indicates whether this item's effect is meant
to be used on all enemies

`Effect`: is the actual effect string

With this change, a function called `get_effect_details()` was made to
get the dictionary for any specific item name. If the item given to
`get_effect_details()` is not in the `Item_Effects.json` file, the
function returns a dictionary of the same form but has empty string for
each value paired to the key.

`get_effect()`: still returns the effect string but does so differently
due to the changes in the `Item_Effects.json` file.

`UI_Handler`: The items local variable is loaded directly from JSON and
some changes were made to process item usage. If the choice the player
made is an item, it get's the effect detail from the `ItemManager`. It
then sets the targets before changing the UI for `UIManager` to "Select
Target". The Helper function `get_enemy_positions()` is now called
instead of manually finding out the enemy positions manually.

`UI_Manager`: The choice that the player made, if it is a string, is
analyzed using regex and if the match is made, the amount portion of the
string is removed and the result tuple is restructured before it is
returned. This is so that "Potion" is passed as the choice instead of
"Potion x1"