Pygame-RPG 17.2 Housekeeping and bugfixing
This PR adds some debugging functionality to the battle system by having
the Entity objects print out their statuses to the console via the
`print_status()` function that. This is then used by the BattleManager
class in the debug function `print_all_statuses()`. This function prints
the status of all active Entities at the end of their turn. This allows
me to have a comprehensive view of what happened during that turn.

A quality of life change was also made to the `take_damage()` and
`take_attack()` functions. Both functions now take in a default value of
`False` at the end, this value relates to whether the damage that will be
taken is from an effect. If the damage is not from an effect, the damage
will be lowered by the object's defence stat. If the damage is from an
effect, this does not happen, making effect essentially true damage.

This PR also fixes a bug that was in the battle system. Due to how the
effect handling was done, the player could never kill enemy units with
the effects. The reason for this is because the changes only affected
the given stat and wouldn't change the status. This way, it was possible
for an enemy object to be below 0 Hp and not get filtered at the end of
the turn.

This is done by ensuring that if the affected stat of an event is Hp,
instead of directly applying the change, `take_damage()` or
`take_attack()` (Depending on the object's class),
is instead called. The boolean flag `True` is also given
to the function so that the damage isn't lowered by the defence stat
(effect damage counts as true damage).

This PR also addresses an oversight that I made when creating the string
array that the battle system gives at the end of every turn. The problem
is that the damage reported is incorrect because it doesn't take into
account the defence of the opponent. Essentially, it'll say the enemy
took 100 damage when in reality their defence reduced it to 30. To fix
this, we have the `take_damage()` and `take_attack()` functions return
an integer that is the actual amount of damage the enemy took. We then
use these for the reporting strings instead. A similar change was done
to the effect version of this situation but it was mostly just to
improve the structure of the code and reduce redundancy.