Pygame-RPG 18.1 Housekeeping and bugfixing
This PR erases some unnecessary comments and adds new event strings to
the `do_one_turn()` functions by adjusting the `clear_dead_enemies()`
function. This function now adds event strings to a list when enemies
die. This PR also refactors the tests to accommodate the changes and
updates the scripts to add a script that adds a status effect to the
Status.json file.