Pygame-RPG 19 Refactor DialogueManager
This PR focuses on refactoring the DialogueManager class. This class now
shows dialogue by having a list of strings as well as 2 "line" string
lists.

As a result of this change in methodology, the `file`, `event` and
`prevText` attributes were erased. The following attributes were added,
`largeFont`, `dialogue`, `backlog`, `firstLine` and `secondLine`.

`largeFont`: This is a font object that is larger than the normal `font`
attribute. This is used exclusively for dialogue in battles.

`dialogue`: This is a list of all the dialogue that needs to be
displayed. This is loaded from a file or a list is directly loaded into
this address.

`backlog`: All the previous dialogue that the player has read. No
implementation for it at the moment.

`firstLine`: A list of all the text that will be displayed on the first
line of the text box

`secondLine`: a list of all the text that will be displayed on the
second line of the text box.

Because of these changes, new functions needed to be added and old
functions needed to be refactored.

`load_file()`: This function received some simplifications because of
the attributes that were removed.

`get_new_text()`: This function removes the first index of the dialogue
list (if the list isn't empty), and adds that into the backlog list.

`load_portrait()`: This function received some simplifications and
received a slight behavioral update. The `name` attribute is only
changed if it has a valid portrait dictionary entry.

`load_dialogue_list()`: This function takes in a list and sets the
`dialogue` attribute to the given list.

`sub_divide_dialogue()`: This function takes in a text string and uses
it to fill the `firstLine` and `secondLine` lists. This function does
this by splitting the string based off a maximum length for a line
string. Once the `firstLine` list is filled, the `secondLine` list begins to
fill. If the `secondLine` list is filled and the text isn't fully
processed, the remaining parts of the text are re-inserted into the
`dialogue` list in index 1. This way, the string is ready to be
reprocessed.

`draw_dialogue()`: This function takes in an eventList and also takes in
a boolean value that indicates if the dialogue is called in a battle.
The only difference between whether the flag is True or False is the
font and the textbox being used. The function uses the `firstLine` and
`secondLine` lists to display the dialogue. Another slight change is
that once the "Enter" key is pressed the `firstLine` and `secondLine`
lists are cleared. The `get_new_text()` function is also
called and if the dialogue list isn't empty, `sub_divide_dialogue()` is
called to refresh the `firstLine` and `secondLine` lists. The texts
received from the `firstLine` and `secondLine` are "cleaned" before
being displayed to the player.

Rpg2.py also received some changes to integrate it into the battle
system. Essentially, the battle cannot continue until the
DialogueManager has no more strings in it's `dialogue` list. The
structure has also been changed so that the state of the battle is only
checked AFTER all the event strings for the turn have been read.

The BattleManager class also received a slight update because it was
adding duplicate event strings. The reason for this is that the code
that adds the event string into the list was part of a while loop
causing duplicates to appear. It also fixed a slight formatting error on
an event string.