Pygame-RPG 19.1 Housekeeping and bugfixing
This PR adds in the functionality that differentiates enemy objects by
their `name` attribute. Since all Enemy objects will be created by the
EntityFactory class, I just added some changes to the class to allow for
this functionality.

I added the `createdCount` attribute and modified the
`create_entity()` function. The `clear_created_count()` function was
also added.

`createdCount`: this attribute is a dictionary that keeps track of the
Entities the factory has created and keeps a count on how many have been
made so far.

`create_entity()`: This function has been slightly modified. Instead of
immediately returning the created object, it creates the object and
references the `createdCount` dictionary. If an entry for the Entity's
name exists, the count mapped to the Entity's name will increase and the
Entity's new name will be their original name plus the count mapped to
the value. If the Entity doesn't have an entry in the dictionary, a new
one is created with the count of 1, it's name is not changed.

`clear_created_count()`: This function just clears the `createdCount`
dictionary.