Pygame-RPG 19.2 Upgrading Battle System
This PR is meant to add new functionality to the parsing of effects.
This PR modifies the `parse_effect()` function as well as the JSON
strings for both item and move effects. This change allows for both
immediate and buff/debuff effects to coexist in one effect string. This
is done by dividing the effects with an indicator. The "|" character in
an effect string indicates the beginning of a new effect.

By splitting the effect string using the "|" character, we are giving
the individual effects in a string form that we can now process. Now,
using this list, we check the length of the array we get by splitting
the list by the "," character and processing them according to their
length (length of 2 is immediate, 3 is buff/debuff).

This next change is designed to imply the target if none is given in the
effect string. While the old format is easy to write, it came with the
downside of always having to write the "S" character as the target
when most of the effects are applied to the user by default

To circumvent this, a new condition is added before the code checks the
length of the array. It now checks the final index of the array to see
if there is a targeting indicator ("S" or "T"). If there isn't, the code
appends "S" to the list making the "S" implied if there is no given
target. To make this change the effect strings format for the
Buff/debuff effect needed to be changed.

Old Format: Buff/Debuff, Target, Duration

New Format: Buff/Debuff, Duration, Target

The only change was that target and duration have swapped positions.
With this new change, the item and move effects were changed to remove
the targeting indicator if the effect only applied to the user and the
effect strings now have the "|" character to divide up the different
effects.

New tests were added to confirm that this feature works including
creating a move with both an immediate and a buff effect to see if it
could be processed.