Pygame-RPG 19.3 Upgrade Battle System
This PR focuses on allowing multi target skills to be usable by the
current battle system. In order to do this, changes to the BattleManager
class and the Rpg2.py file need to be made.

Changes made in BattleManager class:
To track the number of selected targets a move needs, a new attribute
called `targetNum` was created which starts at the value of -1 (meaning
it's inactive). A new function called `set_target_number()` was made and
the `do_one_move()` function was modified.

`targetNum`: This is an integer value that gets set to a positive number
when a move is chosen. Once the move has been used, the value is to be
reset to -1 signifying that it isn't in use.

`set_target_number()`: This function sets the `targetNum` value based
off of the name of the chosen move given.

`do_one_turn()`: The major change made for this function is that the
move of a Knight object is not proccessed unless the amount of targets
matches with the `targetNum` attribute of the class. This way, the
player doesn't use moves prematurely. The player can target the same
enemy multiple times or choose a different one. This change also allows
the enemy to use multi-hit attacks as well, the only difference is that
since the player is the only person that the enemies can target, they'll
simply always target them.

Changes made to the Rpg2.py file:
The target variable that is used for BattleManager is now a list rather
than a tuple. This allows us to add and remove things as we please.
There were also changes relating to the clear conditions of the
`buffered_move` and `target` variables. Now, to clear these variables,
`buffered_move` needs to have a valid value and the length of the
`target` list needs to be equal to the targetNum attribute in the
BattleManager object.

A change was also made when a move is chosen. If the move is a valid
move, the BattleManager classes `targetNum` attribute is set using the
`set_target_number()` function if it is a valid move. If the choice is
an item, the attribute is set to 1 (for now we assume there are no
multi-hit items. The UI also doesn't change if the player hasn't
specified who to hit for every hit of the move.

Finally, there was a quality of life change for when the player selects
a multi-hit move when there's only a single enemy left in the battle.
The game will automatically hit the enemy with all of the hits, saving
the player time.